Added New SliderWithEchoInt script
to support int manipulation for sliders, created a new integer version of the slider with echo. Issue right now: when manipulating the mesh with the UI sliders, tends to error on how many to make and/or how many to delete/recalculate. I hope inclusion of the selected vertex will clear it up